+++ allsecureexchange/woocommerce-allsecure-exchange.php
@@ -3,7 +3,7 @@
  * Plugin Name: WooCommerce AllSecure Exchange Extension
  * Plugin URI: https://help.allsecure.xyz
  * Description: AllSecure Exchange for WooCommerce
- * Version: 1.9.9
+ * Version: 1.9.10
  * Tested up to: 5.8.4
  * Author: AllSecure Exchange Team
  * WC requires at least: 3.6.0
@@ -18,7 +18,7 @@ if (!defined('ABSPATH')) {
 define('ALLSECURE_EXCHANGE_EXTENSION_URL', 'https://asxgw.com/');
 define('ALLSECURE_EXCHANGE_EXTENSION_TEST_URL', 'https://asxgw.paymentsandbox.cloud/');
 define('ALLSECURE_EXCHANGE_EXTENSION_NAME', 'AllSecure Exchange');
-define('ALLSECURE_EXCHANGE_EXTENSION_VERSION', '1.9.9');
+define('ALLSECURE_EXCHANGE_EXTENSION_VERSION', '1.9.10');
 define('ALLSECURE_EXCHANGE_EXTENSION_UID_PREFIX', 'allsecure_exchange_');
 define('ALLSECURE_EXCHANGE_EXTENSION_BASEDIR', plugin_dir_path(__FILE__));